Make TasksetGenerator object
